Move cdnImages.ts with paths to frontpage images to (index) folder and update Account Settings dialog to use logo from Shared WebApp
