+++ application/account-management/WebApp/routes/(index)/-components/FooterSection.tsx
@@ -1,4 +1,4 @@
-import { githubLogo, linkedinLogo, logoWrap, slackLogo, twitterLogo, youtubeLogo } from "@/shared/images/cdnImages";
+import { githubLogo, linkedinLogo, logoWrap, slackLogo, twitterLogo, youtubeLogo } from "./cdnImages";
 import { Button } from "@repo/ui/components/Button";
 import { TextField } from "@repo/ui/components/TextField";
