+++ application/account-management/WebApp/shared/components/accountModals/AccountSettingsModal.tsx
@@ -1,4 +1,4 @@
-import { logoWrap } from "@/shared/images/cdnImages";
+import logoWrap from "@/shared/images/logo-wrap.svg";
 import { Button } from "@repo/ui/components/Button";
 import { Dialog } from "@repo/ui/components/Dialog";
 import { Modal } from "@repo/ui/components/Modal";
